CHANGED DATABASE TABLE FORMAT
oai2sqlv2.py creates an alternate version of the OAI database where
we merge all separate visit data sets into a single table. For example,
JointSx00, JointSx01 ... etc are now in a single table JointSx with
visit information explicitly coded as with the VID field.

This is much nice for querying historical data about a single subject.
Missing fields are encoded as NULLs